backend/requirements.txt
Django==5.0.6
python-slugify==8.0.4
asgiref==3.7.2
django-admin-sortable2==2.1.10
django-debug-toolbar==4.3.0
django-dragndrop-related==0.3.0
django-environ==0.11.2
# mysql==0.0.3
# mysqlclient==2.2.4
pillow==10.2.0
python-dotenv==1.0.1
python-slugify==8.0.4
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2024.1
django-cleanup==8.1.0